[FEAT] basic segmentation pipe + save
