+++ salami/config.py
@@ -6,3 +6,8 @@ import os
 SALAMI_PATH = os.path.dirname(os.path.abspath(__file__))
 PROTOCOL_PATH = os.path.join(SALAMI_PATH, "protocol", "protocol.yaml")
 
+
+
+DATA_DIR = "data"
+SEG_DIR = "seg"
+DENOISE_DIR = "denoise"
