+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = salami
-version = 0.1.0
+version = 0.0.1
 author = Patrick Cleeve
 author_email = Patrick.Cleeve@monash.edu
 description = automated volume electron microscopy
